refactor: moving to helpers.ts
